Restructure directory structure to improve readability. Also, added new functions to reduce code redundancy and further improve readability.
